go mod tidy and fix cast issue for int to int64

+++ influxdbv2/resource_create_buckets.go
@@ -161,7 +161,7 @@ func getRetentionRules(input interface{}) (domain.RetentionRules, error) {
 	for _, retentionRule := range retentionRulesSet {
 		rr, ok := retentionRule.(map[string]interface{})
 		if ok {
-			each := domain.RetentionRule{EverySeconds: rr["every_seconds"].(int64)}
+			each := domain.RetentionRule{EverySeconds: int64(rr["every_seconds"].(int))}
 			result = append(result, each)
 		}
 	}
